Past two sprints worth of work

+++ ios/src/RHITMobileTests/RHBoundaryNodeTests.m
@@ -86,7 +86,7 @@
         storedNode = (RHBoundaryNode *) [results objectAtIndex:0];
         
         // Verify the properties set on our retrieved object
-        STAssertEquals(storedNode.latitude, latitude, @"Latitude is incorrect");
+        STAssertEquals(storedNode.latitude.doubleValue, latitude.doubleValue, @"Latitude is incorrect");
         STAssertEquals(storedNode.longitude.doubleValue, 0.0,
                        @"Longitude is incorrect");
         STAssertEquals(storedNode.position.doubleValue, 4.0,
